pfpki: optimize hot paths, fix latent bugs, split models.go
Hot-path / safety:
- ocspresponder: add a package-level cache for parsed CA cert+key keyed
  by (ID, UpdatedAt). The previous Verify() ran 3 cert parses and 2 key
  parses per OCSP request; now they happen once per CA revision.
- ocspresponder: drop the broken duplicate-nonce tracker. The responder
  is reconstructed per HTTP request (see handlers.Responder), so the
  per-instance NonceList never survived a second request; combined with
  `make([][]byte, 10)` (10 nil entries) it was both a no-op and an
  unbounded-growth foot-gun. The nonce extension is still echoed.
- handlers: new LimitRequestBody chi middleware wraps every request body
  with http.MaxBytesReader (1 MiB). Stops a client from streaming a
  giant body and OOMing the process. Mounted in pfpki.go and testutil.
- models.revokeNeeded: now takes the loaded *Profile from the caller
  instead of fetching pki_profiles itself. Cert.New saves one SELECT
  per leaf-issuance. HasCN now loads the profile once explicitly.
- sql.SqlFields: memoise the reflect.Type → []string mapping with
  sync.Map; previously every list query re-walked the struct tags.

Code-quality:
- Profile.Update: drop the reflect.ValueOf-driven field copier (+ the
  contains/setProperty helpers) and use explicit field assignments.
  Same effect, no reflection, compiler-checked when fields move.
- Profile.New, Cert.New: replace the 40+/20+-field struct literals
  that copied almost every field from the receiver with
  `x := p; x.Foo = …` patterns. Less code, less drift.
- maybeRevokeExpiring: extract the "decode cert PEM, revoke if close
  to expiry" branch shared by both revokeNeeded paths.

Layout:
- models.go (2382 LOC) split by type:
    models.go        (324 LOC) — types, TableName, comments,
                                  dbError, ProfileAttributes
    ca.go            (733 LOC) — CA methods, serial helpers,
                                  revokeNeeded, maybeRevokeExpiring
    profile.go       (217 LOC)
    cert.go          (634 LOC)
    revoked.go        (78 LOC)
    csr.go           (182 LOC)
    email.go         (193 LOC) — EmailType, emailcert, emailRenewal,
                                  email, parseTemplate, ParseYAMLDict,
                                  dictionary, Lookup
    scep_server.go   (116 LOC)
