+++ go/plugin/caddy2/pfpki/models/models.go
@@ -1,56 +1,15 @@
 package models
 
 import (
-	"crypto"
-	"html/template"
-	"net"
-	"net/http"
-	"os"
-	"reflect"
+	"context"
+	"crypto/x509"
+	"strconv"
 	"time"
 
-	"bytes"
-	"crypto/dsa"
-	"crypto/ecdsa"
-	"crypto/ed25519"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/sha1"
-	"crypto/tls"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/hex"
-	"encoding/pem"
-	"errors"
-	"math/big"
-	"strconv"
-	"strings"
-
-	"github.com/inverse-inc/scep/cryptoutil"
-	"github.com/inverse-inc/scep/scep"
-	"github.com/knq/pemutil"
-
-	"context"
-	"fmt"
-	"io"
-
-	"github.com/inverse-inc/go-utils/log"
-	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/certutils"
 	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/cloud"
-	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/sql"
 	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/types"
 	_ "gorm.io/driver/mysql"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
-	pkcs12 "software.sslmate.com/src/go-pkcs12"
-
-	"github.com/inverse-inc/packetfence/go/pfconfigdriver"
-	"golang.org/x/crypto/ocsp"
-	"golang.org/x/text/language"
-	"golang.org/x/text/message"
-	"golang.org/x/text/message/catalog"
-	gomail "gopkg.in/gomail.v2"
-	yaml "gopkg.in/yaml.v2"
 )
 
 type (
@@ -309,1884 +268,6 @@ const dbError = "A database error occured. See log for details."
 // 12 ExtKeyUsageMicrosoftCommercialCodeSigning
 // 13 ExtKeyUsageMicrosoftKernelCodeSigning
 
-// NewCAModel create a CAModel
-func NewCAModel(pfpki *types.Handler) *CA {
-	CA := &CA{}
-
-	CA.DB = pfpki.DB
-	CA.Ctx = *pfpki.Ctx
-
-	return CA
-}
-
-// New create a new CA
-func (c CA) New() (types.Info, error) {
-
-	Information := types.Info{}
-
-	keyOut, pub, key, err := certutils.GenerateKey(*c.KeyType, c.KeySize)
-
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	skid, err := certutils.CalculateSKID(pub)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	var cadb CA
-
-	var SerialNumber *big.Int
-
-	if CaDB := c.DB.Last(&cadb); CaDB.Error != nil {
-		SerialNumber = big.NewInt(1)
-	} else {
-		SerialNumber = big.NewInt(int64(cadb.ID + 1))
-	}
-
-	Subject := c.MakeSubject()
-
-	ca := &x509.Certificate{
-		SerialNumber:          SerialNumber,
-		Subject:               Subject,
-		NotBefore:             time.Now(),
-		NotAfter:              time.Now().AddDate(0, 0, c.Days),
-		IsCA:                  true,
-		SignatureAlgorithm:    certutils.CompatibleSigAlgo(*c.KeyType, c.Digest),
-		ExtKeyUsage:           certutils.Extkeyusage(strings.Split(*c.ExtendedKeyUsage, "|")),
-		KeyUsage:              x509.KeyUsage(certutils.Keyusage(strings.Split(*c.KeyUsage, "|"))),
-		BasicConstraintsValid: true,
-		EmailAddresses:        []string{c.Mail},
-		SubjectKeyId:          skid,
-		AuthorityKeyId:        skid,
-	}
-
-	if len(c.OCSPUrl) > 0 {
-		ca.OCSPServer = []string{c.OCSPUrl}
-	}
-
-	var caBytes []byte
-
-	switch *c.KeyType {
-	case certutils.KEY_RSA:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(*rsa.PrivateKey))
-	case certutils.KEY_ECDSA:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(*ecdsa.PrivateKey))
-	case certutils.KEY_DSA:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(*dsa.PrivateKey))
-	case certutils.KEY_ED25519:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(ed25519.PrivateKey))
-	}
-	if err != nil {
-		return Information, err
-	}
-
-	cert := new(bytes.Buffer)
-
-	// Public key
-	pem.Encode(cert, &pem.Block{Type: "CERTIFICATE", Bytes: caBytes})
-
-	// Calculate the IssuerNameHash
-	catls, err := tls.X509KeyPair([]byte(cert.String()), []byte(keyOut.String()))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	h := sha1.New()
-
-	h.Write(cacert.RawIssuer)
-
-	newCA := CA{Cn: c.Cn, Mail: c.Mail, Organisation: c.Organisation, OrganisationalUnit: c.OrganisationalUnit, Country: c.Country, State: c.State, Locality: c.Locality, StreetAddress: c.StreetAddress, PostalCode: c.PostalCode, KeyType: c.KeyType, KeySize: c.KeySize, Digest: c.Digest, KeyUsage: c.KeyUsage, ExtendedKeyUsage: c.ExtendedKeyUsage, Days: c.Days, Key: keyOut.String(), Cert: cert.String(), IssuerKeyHash: hex.EncodeToString(skid), IssuerNameHash: hex.EncodeToString(h.Sum(nil)), OCSPUrl: c.OCSPUrl, SerialNumber: 1}
-	if err := c.DB.Create(&newCA).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-
-	Information.Entries = []CA{newCA}
-
-	return Information, nil
-}
-
-func (c CA) MakeSubject() pkix.Name {
-	var Subject pkix.Name
-	Subject.CommonName = c.Cn
-
-	if len(c.Organisation) > 0 {
-		Subject.Organization = []string{c.Organisation}
-	}
-
-	if len(c.OrganisationalUnit) > 0 {
-		Subject.OrganizationalUnit = []string{c.OrganisationalUnit}
-	}
-
-	if len(c.Country) > 0 {
-		Subject.Country = []string{c.Country}
-	}
-
-	if len(c.State) > 0 {
-		Subject.Province = []string{c.State}
-	}
-
-	if len(c.Locality) > 0 {
-		Subject.Locality = []string{c.Locality}
-	}
-
-	if len(c.StreetAddress) > 0 {
-		Subject.StreetAddress = []string{c.StreetAddress}
-	}
-
-	if len(c.PostalCode) > 0 {
-		Subject.PostalCode = []string{c.PostalCode}
-	}
-	return Subject
-}
-
-// GetByID retreive the CA by id
-func (c CA) GetByID(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var cadb []CA
-	if val, ok := params["id"]; ok {
-		allFields := strings.Join(sql.SqlFields(c)[:], ",")
-		c.DB.Select(allFields).Where("`id` = ?", val).First(&cadb)
-	}
-	Information.Entries = cadb
-
-	return Information, nil
-}
-
-// Fix calculate the IssuerKeyHash and IssuerNameHash
-func (c CA) Fix() (types.Info, error) {
-	Information := types.Info{}
-	var cadb []CA
-
-	c.DB.Find(&cadb)
-	for _, v := range cadb {
-		if v.IssuerNameHash == "" {
-
-			// Calculate the IssuerNameHash
-			catls, err := tls.X509KeyPair([]byte(v.Cert), []byte(v.Key))
-			if err != nil {
-				Information.Error = err.Error()
-				return Information, err
-			}
-			cacert, err := x509.ParseCertificate(catls.Certificate[0])
-			if err != nil {
-				Information.Error = err.Error()
-				return Information, err
-			}
-			h := sha1.New()
-
-			h.Write(cacert.RawIssuer)
-			// var store pemutil.Store
-			store := make(map[pemutil.BlockType]interface{})
-
-			pemutil.Decode(store, []byte(v.Cert))
-			var skid []byte
-			for _, pemUtil := range store {
-				cert := pemUtil.(*x509.Certificate)
-				skid, _ = certutils.CalculateSKID(cert.PublicKey)
-			}
-
-			v.IssuerKeyHash = hex.EncodeToString(skid)
-			v.IssuerNameHash = hex.EncodeToString(h.Sum(nil))
-			c.DB.Save(&v)
-		}
-	}
-
-	return Information, nil
-}
-
-// Paginated return the CA list paginated
-func (c CA) Paginated(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	var count int64
-	c.DB.Model(&CA{}).Count(&count)
-	counter := int(count)
-
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		sql, err := vars.Sql(c)
-		if err != nil {
-			Information.Error = err.Error()
-			return Information, errors.New(dbError)
-		}
-		var cadb []CA
-		c.DB.Select(sql.Select).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&cadb)
-		Information.Entries = cadb
-	}
-
-	return Information, nil
-}
-
-// Search for the CA
-func (c CA) Search(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	sql, err := vars.Sql(c)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-	var count int64
-	c.DB.Model(&CA{}).Where(sql.Where.Query, sql.Where.Values...).Count(&count)
-	counter := int(count)
-
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		var cadb []CA
-		c.DB.Select(sql.Select).Where(sql.Where.Query, sql.Where.Values...).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&cadb)
-		Information.Entries = cadb
-	}
-
-	return Information, nil
-}
-
-// FindSCEPProfile search the SCEP Profile by the profile name
-func (c *CA) FindSCEPProfile(options []string) ([]Profile, error) {
-	var profiledb []Profile
-	profile := &Profile{}
-	if len(options) >= 1 {
-		if err := c.DB.Preload("ScepServer").Where("name = ? and `scep_enabled` = ?", options[0], "1").First(profile).Error; err != nil {
-			if errors.Is(err, gorm.ErrRecordNotFound) {
-				return profiledb, errors.New("Unknow profile.")
-			}
-			return profiledb, errors.New(dbError)
-		}
-		profiledb = append(profiledb, *profile)
-	} else {
-		c.DB.Preload("ScepServer").Where("`scep_enabled` = ?", "1").First(&profiledb)
-	}
-	c.SCEPAssociateProfile = profiledb[0].Name
-
-	return profiledb, nil
-
-}
-
-// CA return the CA public key based on the profile name (SCEP)
-func (c CA) CA(pass []byte, options ...string) ([]*x509.Certificate, *rsa.PrivateKey, error) {
-
-	var profiledb []Profile
-
-	profiledb, err := c.FindSCEPProfile(options)
-
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Proxy the request if a SCEPServer is defined in the profil
-
-	var ca CA
-
-	if err := c.DB.First(&ca, profiledb[0].CaID).Error; err != nil {
-		c.DB.First(&ca)
-	}
-
-	catls, err := tls.X509KeyPair([]byte(ca.Cert), []byte(ca.Key))
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	key, err := certutils.LoadKey([]byte(ca.Key), pass)
-	return []*x509.Certificate{cacert}, key, err
-}
-
-// CA return the CA public key based on the profile name (SCEP)
-func (c CA) CAbyProfile(pass []byte, profilename string) ([]*x509.Certificate, *rsa.PrivateKey, error) {
-	profile := &Profile{}
-	if err := c.DB.Where("name = ?", profilename).First(profile).Error; err != nil {
-		return nil, nil, err
-	}
-
-	var ca CA
-
-	if err := c.DB.First(&ca, profile.CaID).Error; err != nil {
-		c.DB.First(&ca)
-	}
-
-	catls, err := tls.X509KeyPair([]byte(ca.Cert), []byte(ca.Key))
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	key, err := certutils.LoadKey([]byte(ca.Key), pass)
-	return []*x509.Certificate{cacert}, key, err
-}
-
-// Put create the public key in the DB (SCEP)
-func (c CA) Put(cn string, crt *x509.Certificate, options ...string) error {
-
-	attributeMap := certutils.GetDNFromCert(crt.Subject)
-
-	publicKey := new(bytes.Buffer)
-
-	pem.Encode(publicKey, &pem.Block{Type: "CERTIFICATE", Bytes: crt.Raw})
-
-	profiledb, err := c.FindSCEPProfile(options)
-
-	if err != nil {
-		return err
-	}
-
-	var ca CA
-
-	if err := c.DB.First(&ca, profiledb[0].CaID).Error; err != nil {
-		c.DB.First(&ca)
-	}
-	notFalse := true
-	var IPAddresses []string
-	for _, IP := range crt.IPAddresses {
-		IPAddresses = append(IPAddresses, IP.String())
-	}
-
-	if err := c.DB.Create(&Cert{Cn: cn, Ca: ca, CaName: ca.Cn, ProfileName: profiledb[0].Name, SerialNumber: crt.SerialNumber.String(), Mail: strings.Join(crt.EmailAddresses, ","), StreetAddress: attributeMap["streetAddress"], Organisation: attributeMap["O"], OrganisationalUnit: attributeMap["OU"], Country: attributeMap["C"], State: attributeMap["ST"], Locality: attributeMap["L"], PostalCode: attributeMap["postalCode"], DNSNames: strings.Join(crt.DNSNames, ","), IPAddresses: strings.Join(IPAddresses, ","), Profile: profiledb[0], Key: "", Cert: publicKey.String(), ValidUntil: crt.NotAfter, NotBefore: crt.NotBefore, Scep: &notFalse, Subject: crt.Subject.String()}).Error; err != nil {
-		return errors.New(dbError)
-	}
-
-	return nil
-}
-
-// Serial return the serial number
-func (c CA) Serial(options ...string) (*big.Int, error) {
-
-	profiledb, err := c.FindSCEPProfile(options)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return c.FindSerial(profiledb[0])
-}
-
-// FindSerial atomically allocates the next serial number for a Profile's CA.
-// If the profile's CA id can't be found, falls back to the first CA row.
-func (c CA) FindSerial(p Profile) (*big.Int, error) {
-	return nextSerialNumber(c.DB, p.CaID, true)
-}
-
-// getNextSerialNumber atomically increments and returns the next serial
-// number for the given CA. Uses a row-locked transaction to prevent race
-// conditions.
-func getNextSerialNumber(db *gorm.DB, caID uint) (*big.Int, error) {
-	return nextSerialNumber(db, caID, false)
-}
-
-// nextSerialNumber holds the shared row-locked increment used by both
-// FindSerial (with a fallback to the first CA) and getNextSerialNumber.
-func nextSerialNumber(db *gorm.DB, caID uint, fallbackToFirst bool) (*big.Int, error) {
-	var serialNumber int
-
-	err := db.Transaction(func(tx *gorm.DB) error {
-		ca := &CA{}
-		locked := tx.Clauses(clause.Locking{Strength: "UPDATE"})
-		if err := locked.First(ca, caID).Error; err != nil {
-			if !fallbackToFirst {
-				return err
-			}
-			if err := locked.First(ca).Error; err != nil {
-				return err
-			}
-		}
-
-		serialNumber = ca.SerialNumber
-		ca.SerialNumber = ca.SerialNumber + 1
-		return tx.Save(ca).Error
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return big.NewInt(int64(serialNumber)), nil
-}
-
-func (c CA) HasCN(cn string, allowTime int, cert *x509.Certificate, revokeOldCertificate bool, options ...string) (bool, error) {
-
-	return revokeNeeded(cn, options[0], allowTime, c.DB)
-
-}
-
-func revokeNeeded(cn string, profile string, allowTime int, c *gorm.DB) (bool, error) {
-
-	var certif Cert
-	var certifs []Cert
-	var CertDB *gorm.DB
-
-	var profiledb []Profile
-	c.Where("name = ?", profile).First(&profiledb)
-	if profiledb[0].AllowDuplicatedCN == 1 {
-		// Allow duplicated CN in the DB for this profile
-		if profiledb[0].MaximumDuplicatedCN == 0 {
-			return true, nil
-		}
-		if CertDB = c.Where("cn = ? AND profile_name = ?", cn, profile).Find(&certifs); CertDB.Error == nil {
-			// Do we have to revoke some of them ?
-			i := 0
-			for _, certificat := range certifs {
-				certificat.DB = c
-
-				store := make(map[pemutil.BlockType]interface{})
-
-				pemutil.Decode(store, []byte(certificat.Cert))
-				for _, pemUtil := range store {
-					cert := pemUtil.(*x509.Certificate)
-					if cert.NotAfter.Unix()-int64((time.Duration(allowTime)*24*time.Hour).Seconds()) < time.Now().Unix() || allowTime == 0 {
-
-						params := make(map[string]string)
-
-						params["id"] = strconv.Itoa(int(certificat.ID))
-						params["reason"] = strconv.Itoa(ocsp.Superseded)
-						certificat.Revoke(params)
-						i++
-					}
-				}
-			}
-			if int(CertDB.RowsAffected)-i >= profiledb[0].MaximumDuplicatedCN {
-				return false, errors.New("Certificate with this Subject already exist: " + cn)
-			}
-			// There is no certificate with this CN in the DB
-			return true, nil
-		}
-	}
-	if CertDB = c.Where("Cn = ? AND profile_name = ?", cn, profile).Find(&certif); CertDB.Error != nil {
-		// There is no certificate with this CN in the DB
-		return true, nil
-	}
-
-	if CertDB.RowsAffected == 0 {
-		return true, nil
-	}
-	certif.DB = c
-
-	store := make(map[pemutil.BlockType]interface{})
-	pemutil.Decode(store, []byte(certif.Cert))
-	for _, pemUtil := range store {
-		cert := pemUtil.(*x509.Certificate)
-		if cert.NotAfter.Unix()-int64((time.Duration(allowTime)*24*time.Hour).Seconds()) < time.Now().Unix() || allowTime == 0 {
-
-			params := make(map[string]string)
-
-			params["id"] = strconv.Itoa(int(certif.ID))
-			params["reason"] = strconv.Itoa(ocsp.Superseded)
-			certif.Revoke(params)
-			return true, nil
-		}
-	}
-
-	return false, errors.New("Certificate with this Subject already exist: " + cn)
-
-}
-
-// SCEP Verify
-func (c CA) Verify(m *scep.CSRReqMessage) (bool, error) {
-	prof, _ := c.GetProfileByName(c.SCEPAssociateProfile)
-
-	if prof.CloudEnabled == 1 {
-		vcloud, err := cloud.Create(c.Ctx, "intune", prof.CloudService)
-		if err != nil {
-			return false, err
-		}
-		err = vcloud.ValidateRequest(c.Ctx, m.CSR.Raw)
-
-		if err != nil {
-			return false, err
-		}
-		c.Cloud = vcloud
-		return true, nil
-	}
-	return true, nil
-}
-
-func (c CA) FailureNotify(cert *x509.Certificate, m *scep.CSRReqMessage, message string) {
-	if c.Cloud != nil {
-		c.Cloud.FailureReply(c.Ctx, cert, m.CSR.Raw, message)
-	}
-}
-
-func (c CA) SuccessNotify(cert *x509.Certificate, m *scep.CSRReqMessage, message string) {
-	if c.Cloud != nil {
-		c.Cloud.SuccessReply(c.Ctx, cert, m.CSR.Raw, message)
-	}
-}
-
-func (c CA) GetProfileByName(name string) (*Profile, error) {
-	var profiledb []Profile
-	c.DB.Where("name = ?", name).First(&profiledb)
-
-	return &profiledb[0], nil
-}
-
-func (c CA) Resign(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var cadb []CA
-	var err error
-	if val, ok := params["id"]; ok {
-		if err = c.DB.First(&cadb, val).Error; err != nil {
-			Information.Error = err.Error()
-			return Information, err
-		}
-
-	}
-
-	Information.Entries = cadb
-
-	block, _ := pem.Decode([]byte(cadb[0].Key))
-	if block == nil {
-		log.LoggerWContext(c.Ctx).Error("failed to decode PEM block containing public key")
-	}
-
-	var skid []byte
-	var keyOut *bytes.Buffer
-	keyOut = new(bytes.Buffer)
-	var key crypto.PrivateKey
-	var pub crypto.PublicKey
-	keyOut, skid, pub, key, Information, err = certutils.ExtractPrivateKey(c.KeyType, block, &Information)
-	if err != nil {
-		return Information, err
-	}
-
-	var cadbprevious CA
-	var newcadb []CA
-
-	var SerialNumber *big.Int
-
-	if CaDB := c.DB.Last(&cadbprevious); CaDB.Error != nil {
-		SerialNumber = big.NewInt(1)
-	} else {
-		SerialNumber = big.NewInt(int64(cadbprevious.ID + 1))
-	}
-
-	Subject := c.MakeSubject()
-
-	ca := &x509.Certificate{
-		SerialNumber:          SerialNumber,
-		Subject:               Subject,
-		NotBefore:             time.Now(),
-		NotAfter:              time.Now().AddDate(0, 0, c.Days),
-		IsCA:                  true,
-		SignatureAlgorithm:    certutils.CompatibleSigAlgo(*c.KeyType, c.Digest),
-		ExtKeyUsage:           certutils.Extkeyusage(strings.Split(*c.ExtendedKeyUsage, "|")),
-		KeyUsage:              x509.KeyUsage(certutils.Keyusage(strings.Split(*c.KeyUsage, "|"))),
-		BasicConstraintsValid: true,
-		EmailAddresses:        []string{c.Mail},
-		SubjectKeyId:          skid,
-		AuthorityKeyId:        skid,
-	}
-
-	if len(c.OCSPUrl) > 0 {
-		ca.OCSPServer = []string{c.OCSPUrl}
-	}
-
-	var caBytes []byte
-
-	switch *c.KeyType {
-	case certutils.KEY_RSA:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(*rsa.PrivateKey))
-	case certutils.KEY_ECDSA:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(*ecdsa.PrivateKey))
-	case certutils.KEY_DSA:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(*dsa.PrivateKey))
-	case certutils.KEY_ED25519:
-		caBytes, err = x509.CreateCertificate(rand.Reader, ca, ca, pub, key.(ed25519.PrivateKey))
-	}
-	if err != nil {
-		return Information, err
-	}
-
-	cert := new(bytes.Buffer)
-	// Public key
-	pem.Encode(cert, &pem.Block{Type: "CERTIFICATE", Bytes: caBytes})
-
-	// Calculate the IssuerNameHash
-	catls, err := tls.X509KeyPair([]byte(cert.String()), []byte(keyOut.String()))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	h := sha1.New()
-
-	h.Write(cacert.RawIssuer)
-
-	if err := c.DB.Model(&CA{}).Where("cn = ?", c.Cn).Updates(map[string]interface{}{"Cn": c.Cn, "Mail": c.Mail, "Organisation": c.Organisation, "OrganisationalUnit": c.OrganisationalUnit, "Country": c.Country, "State": c.State, "Locality": c.Locality, "StreetAddress": c.StreetAddress, "PostalCode": c.PostalCode, "KeyType": c.KeyType, "KeySize": c.KeySize, "Digest": c.Digest, "KeyUsage": c.KeyUsage, "ExtendedKeyUsage": c.ExtendedKeyUsage, "Days": c.Days, "Key": keyOut.String(), "Cert": cert.String(), "IssuerKeyHash": hex.EncodeToString(skid), "IssuerNameHash": hex.EncodeToString(h.Sum(nil)), "OCSPUrl": c.OCSPUrl}).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New("A database error occured. See log for details.")
-	}
-
-	c.DB.Where("cn = ?", c.Cn).First(&newcadb)
-	Information.Entries = newcadb
-
-	return Information, nil
-}
-
-func (c CA) GenerateCSR(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var cadb []CA
-	var err error
-	if val, ok := params["id"]; ok {
-		if err = c.DB.First(&cadb, val).Error; err != nil {
-			Information.Status = http.StatusNotFound
-			Information.Error = err.Error()
-			return Information, err
-		}
-
-	}
-	catls, err := tls.X509KeyPair([]byte(cadb[0].Cert), []byte(cadb[0].Key))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	Information.Entries = cadb
-
-	// Use the stored CA's KeyType/Digest, not the request body's — the
-	// receiver may only carry the subject fields the caller wants to change,
-	// and the signing algorithm is bound to the existing private key anyway.
-	template := x509.CertificateRequest{
-		Subject:            c.MakeSubject(),
-		SignatureAlgorithm: certutils.CompatibleSigAlgo(*cadb[0].KeyType, cadb[0].Digest),
-	}
-	csrBuff := new(bytes.Buffer)
-	csrBytes, err := x509.CreateCertificateRequest(rand.Reader, &template, catls.PrivateKey)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	pem.Encode(csrBuff, &pem.Block{Type: "CERTIFICATE REQUEST", Bytes: csrBytes})
-	Information.Entries = csrBuff.String()
-
-	if err := c.DB.Model(&CA{}).Where("cn = ?", c.Cn).Updates(map[string]interface{}{"Cn": c.Cn, "Mail": c.Mail, "Organisation": c.Organisation, "OrganisationalUnit": c.OrganisationalUnit, "Country": c.Country, "State": c.State, "Locality": c.Locality, "StreetAddress": c.StreetAddress, "PostalCode": c.PostalCode, "KeyType": c.KeyType, "KeySize": c.KeySize, "Digest": c.Digest, "KeyUsage": c.KeyUsage, "ExtendedKeyUsage": c.ExtendedKeyUsage, "Days": c.Days, "OCSPUrl": c.OCSPUrl}).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New("A database error occured. See log for details.")
-	}
-
-	return Information, err
-
-}
-
-func (c CA) Update(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var cadb []CA
-	var err error
-	if val, ok := params["id"]; ok {
-		if err = c.DB.First(&cadb, val).Error; err != nil {
-			Information.Status = http.StatusNotFound
-			Information.Error = err.Error()
-			return Information, err
-		}
-	}
-	_, err = tls.X509KeyPair([]byte(c.Cert), []byte(cadb[0].Key))
-
-	if err != nil {
-		Information.Error = err.Error()
-		Information.Status = http.StatusUnprocessableEntity
-		return Information, nil
-	}
-	cadb[0].Cert = c.Cert
-	c.DB.Save(&cadb[0])
-	Information.Entries = cadb[0]
-	return Information, err
-}
-
-// EST
-func (c CA) CACerts(ctx context.Context, aps string, r *http.Request) ([]*x509.Certificate, error) {
-	var certs []*x509.Certificate
-	catls, err := tls.X509KeyPair([]byte(c.Cert), []byte(c.Key))
-	if err != nil {
-		log.LoggerWContext(c.Ctx).Error(err.Error())
-	}
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	if err != nil {
-		log.LoggerWContext(c.Ctx).Error(err.Error())
-	}
-	certs = append(certs, cacert)
-	return certs, nil
-}
-
-// Profile section
-func NewProfileModel(pfpki *types.Handler) *Profile {
-	Profile := &Profile{}
-
-	Profile.DB = pfpki.DB
-	Profile.Ctx = *pfpki.Ctx
-
-	return Profile
-}
-
-func (p Profile) New() (types.Info, error) {
-
-	var err error
-	Information := types.Info{}
-	switch *p.KeyType {
-	case certutils.KEY_RSA:
-		if p.KeySize < 2048 {
-			err = errors.New("invalid private key size, should be at least 2048")
-			Information.Error = err.Error()
-			return Information, err
-		}
-	case certutils.KEY_ECDSA:
-		if !(p.KeySize == 256 || p.KeySize == 384 || p.KeySize == 521) {
-			err = errors.New("invalid private key size, should be 256 or 384 or 521")
-			Information.Error = err.Error()
-			return Information, err
-		}
-	case certutils.KEY_DSA:
-		if !(p.KeySize == 1024 || p.KeySize == 2048 || p.KeySize == 3072) {
-			err = errors.New("invalid private key size, should be 1024 or 2048 or 3072")
-			Information.Error = err.Error()
-			return Information, err
-		}
-	case certutils.KEY_ED25519:
-		// Ed25519 has a fixed key size; ignore p.KeySize.
-	default:
-		return Information, errors.New("KeyType unsupported")
-
-	}
-
-	ca := &CA{}
-
-	if err := p.DB.First(ca, p.CaID).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	scepserver := &SCEPServer{}
-	// Choose the default scep server in the db
-	if p.ScepServerID == 0 {
-		p.ScepServerID = 1
-	}
-
-	if err := p.DB.First(scepserver, p.ScepServerID).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	profile := Profile{Name: p.Name, Ca: *ca, CaID: p.CaID, CaName: ca.Cn, Mail: p.Mail, StreetAddress: p.StreetAddress, Organisation: p.Organisation, OrganisationalUnit: p.OrganisationalUnit, Country: p.Country, State: p.State, Locality: p.Locality, PostalCode: p.PostalCode, Validity: p.Validity, KeyType: p.KeyType, KeySize: p.KeySize, Digest: p.Digest, KeyUsage: p.KeyUsage, ExtendedKeyUsage: p.ExtendedKeyUsage, OCSPUrl: p.OCSPUrl, P12MailPassword: p.P12MailPassword, P12MailSubject: p.P12MailSubject, P12MailFrom: p.P12MailFrom, P12MailHeader: p.P12MailHeader, P12MailFooter: p.P12MailFooter, SCEPEnabled: p.SCEPEnabled, SCEPChallengePassword: p.SCEPChallengePassword, SCEPDaysBeforeRenewal: p.SCEPDaysBeforeRenewal, DaysBeforeRenewal: p.DaysBeforeRenewal, RenewalMail: p.RenewalMail, DaysBeforeRenewalMail: p.DaysBeforeRenewalMail, RenewalMailSubject: p.RenewalMailSubject, RenewalMailFrom: p.RenewalMailFrom, RenewalMailHeader: p.RenewalMailHeader, RenewalMailFooter: p.RenewalMailFooter, RevokedValidUntil: p.RevokedValidUntil, CloudEnabled: p.CloudEnabled, CloudService: p.CloudService, ScepServerEnabled: p.ScepServerEnabled, ScepServerID: p.ScepServerID, ScepServer: p.ScepServer, AllowDuplicatedCN: p.AllowDuplicatedCN, MaximumDuplicatedCN: p.MaximumDuplicatedCN}
-
-	if err := p.DB.Create(&profile).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-	Information.Entries = []Profile{profile}
-
-	return Information, nil
-}
-
-func (p Profile) Update(params map[string]string) (types.Info, error) {
-	var profiledb []Profile
-	Information := types.Info{}
-	scepserver := &SCEPServer{}
-
-	profile := &Profile{}
-
-	// Choose the default scep server in the db
-	if p.ScepServerID == 0 {
-		p.ScepServerID = 1
-	}
-
-	if err := p.DB.First(scepserver, p.ScepServerID).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	if val, ok := params["id"]; ok {
-		if err := p.DB.First(profile, val).Error; err != nil {
-			Information.Error = err.Error()
-			return Information, err
-		}
-	} else {
-		if err := p.DB.Where("name = ?", p.Name).First(profile).Error; err != nil {
-			Information.Error = err.Error()
-			return Information, err
-		}
-	}
-
-	fieldsToExtract := []string{"Mail", "Organisation", "OrganisationalUnit", "Country", "State", "Locality", "StreetAddress", "PostalCode", "Validity", "KeyUsage", "ExtendedKeyUsage", "OCSPUrl", "P12MailPassword", "P12MailSubject", "P12MailFrom", "P12MailHeader", "P12MailFooter", "SCEPEnabled", "SCEPChallengePassword", "SCEPDaysBeforeRenewal", "DaysBeforeRenewal", "RenewalMail", "DaysBeforeRenewalMail", "RenewalMailSubject", "RenewalMailFrom", "RenewalMailHeader", "RenewalMailFooter", "RevokedValidUntil", "CloudEnabled", "CloudService", "ScepServerEnabled", "AllowDuplicatedCN", "MaximumDuplicatedCN"}
-
-	v := reflect.ValueOf(p)
-	typeOfS := v.Type()
-	for i := 0; i < v.NumField(); i++ {
-		if contains(fieldsToExtract, typeOfS.Field(i).Name) {
-			profile = profile.setProperty(typeOfS.Field(i).Name, v.Field(i).Interface())
-		}
-	}
-
-	profile.ScepServer = *scepserver
-	profile.SCEPEnabled = p.SCEPEnabled
-
-	p.DB.Save(&profile)
-
-	profiledb = append(profiledb, p)
-	Information.Entries = profiledb
-
-	return Information, nil
-}
-
-func (p *Profile) setProperty(propName string, propValue interface{}) *Profile {
-	reflect.ValueOf(p).Elem().FieldByName(propName).Set(reflect.ValueOf(propValue))
-	return p
-}
-
-func contains(s []string, str string) bool {
-	for _, v := range s {
-		if v == str {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (p Profile) GetByID(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var profiledb []Profile
-	if val, ok := params["id"]; ok {
-		allFields := strings.Join(sql.SqlFields(p)[:], ",")
-		p.DB.Select(allFields).Where("`id` = ?", val).First(&profiledb)
-	}
-	Information.Entries = profiledb
-
-	return Information, nil
-}
-
-func (p Profile) Paginated(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	var count int64
-	p.DB.Model(&Profile{}).Count(&count)
-	counter := int(count)
-
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		sql, err := vars.Sql(p)
-		if err != nil {
-			Information.Error = err.Error()
-			return Information, errors.New(dbError)
-		}
-		var profiledb []Profile
-		p.DB.Select(sql.Select).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&profiledb)
-		Information.Entries = profiledb
-	}
-
-	return Information, nil
-}
-
-func (p Profile) Search(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	sql, err := vars.Sql(p)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-	var count int64
-	p.DB.Model(&Profile{}).Where(sql.Where.Query, sql.Where.Values...).Count(&count)
-	counter := int(count)
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		var profiledb []Profile
-		p.DB.Select(sql.Select).Where(sql.Where.Query, sql.Where.Values...).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&profiledb)
-		Information.Entries = profiledb
-	}
-
-	return Information, nil
-}
-
-func NewCertModel(pfpki *types.Handler) *Cert {
-	Cert := &Cert{}
-
-	Cert.DB = pfpki.DB
-	Cert.Ctx = *pfpki.Ctx
-
-	return Cert
-}
-
-func (c Cert) New() (types.Info, error) {
-	Information := types.Info{}
-	Information.Status = http.StatusUnprocessableEntity
-	// Find the profile
-	var prof Profile
-	if profDB := c.DB.Preload("Ca").First(&prof, c.ProfileID); profDB.Error != nil {
-		Information.Error = profDB.Error.Error()
-		return Information, errors.New(dbError)
-	}
-
-	// Check if the certificate is allowed to be revoked
-	_, err := revokeNeeded(c.Cn, prof.Name, prof.DaysBeforeRenewal, c.DB)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	// Load the certificates from the database
-	catls, err := tls.X509KeyPair([]byte(prof.Ca.Cert), []byte(prof.Ca.Key))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	var SerialNumber *big.Int
-
-	// Get serial number atomically using transaction with row-level locking
-	SerialNumber, err = getNextSerialNumber(c.DB, prof.Ca.ID)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	keyOut, pub, _, err := certutils.GenerateKey(*prof.KeyType, prof.KeySize)
-
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	skid, err := certutils.CalculateSKID(pub)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	Subject := c.MakeSubject()
-
-	NotAfter := c.ValidUntil
-	if c.ValidUntil.IsZero() {
-		NotAfter = time.Now().AddDate(0, 0, prof.Validity)
-	}
-
-	// Prepare certificate
-	cert := &x509.Certificate{
-		SerialNumber:       SerialNumber,
-		Subject:            Subject,
-		NotBefore:          time.Now(),
-		NotAfter:           NotAfter,
-		SignatureAlgorithm: certutils.CompatibleSigAlgo(*prof.Ca.KeyType, prof.Digest),
-		ExtKeyUsage:        certutils.Extkeyusage(strings.Split(*prof.ExtendedKeyUsage, "|")),
-		KeyUsage:           x509.KeyUsage(certutils.Keyusage(strings.Split(*prof.KeyUsage, "|"))),
-		SubjectKeyId:       skid,
-	}
-
-	if len(prof.OCSPUrl) > 0 {
-		cert.OCSPServer = []string{prof.OCSPUrl}
-	}
-
-	Email := ""
-	if len(prof.Mail) > 0 {
-		Email = prof.Mail
-	}
-	if len(c.Mail) > 0 {
-		Email = c.Mail
-	}
-
-	if len(Email) > 0 {
-		for _, mail := range strings.Split(Email, ",") {
-			cert.EmailAddresses = append(cert.EmailAddresses, mail)
-		}
-	}
-
-	if len(c.DNSNames) > 0 {
-		for _, dns := range strings.Split(c.DNSNames, ",") {
-			cert.DNSNames = append(cert.DNSNames, dns)
-		}
-	}
-	var IPAddresses []string
-	if len(c.IPAddresses) > 0 {
-		for _, ip := range strings.Split(c.IPAddresses, ",") {
-			if net.ParseIP(ip) == nil {
-				fmt.Printf("IP Address: %s - Invalid\n", ip)
-			} else {
-				IPAddresses = append(IPAddresses, ip)
-				cert.IPAddresses = append(cert.IPAddresses, net.ParseIP(ip))
-			}
-		}
-	}
-
-	// Sign the certificate
-	certByte, err := x509.CreateCertificate(rand.Reader, cert, cacert, pub, catls.PrivateKey)
-
-	certBuff := new(bytes.Buffer)
-
-	// Public key
-	pem.Encode(certBuff, &pem.Block{Type: "CERTIFICATE", Bytes: certByte})
-
-	newCert := Cert{Cn: c.Cn, Ca: prof.Ca, CaName: prof.Ca.Cn, ProfileName: prof.Name, SerialNumber: SerialNumber.String(), DNSNames: c.DNSNames, IPAddresses: strings.Join(IPAddresses, ","), Mail: Email, StreetAddress: strings.Join(Subject.StreetAddress, ""), Organisation: strings.Join(Subject.Organization, ""), OrganisationalUnit: strings.Join(Subject.OrganizationalUnit, ""), Country: strings.Join(Subject.Country, ""), State: strings.Join(Subject.Province, ""), Locality: strings.Join(Subject.Locality, ""), PostalCode: strings.Join(Subject.PostalCode, ""), Profile: prof, Key: keyOut.String(), Cert: certBuff.String(), ValidUntil: cert.NotAfter, NotBefore: cert.NotBefore, Subject: Subject.String()}
-	if err := c.DB.Create(&newCert).Error; err != nil {
-		Information.Error = err.Error()
-		Information.Status = http.StatusConflict
-		return Information, errors.New(dbError)
-	}
-	log.LoggerWContext(c.Ctx).Info("Certificate " + c.Cn + " has been generated from profile " + prof.Name + " and sign by " + prof.Ca.Cn)
-	Information.Entries = []Cert{newCert}
-	Information.Serial = SerialNumber.String()
-
-	return Information, nil
-}
-
-func (c Cert) GetByID(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var certdb []Cert
-	allFields := strings.Join(sql.SqlFields(c)[:], ",")
-	if val, ok := params["id"]; ok {
-		c.DB.Select(allFields).Where("`id` = ?", val).First(&certdb)
-	}
-	if val, ok := params["cn"]; ok {
-		c.DB.Select(allFields).Where("`cn` = ?", val).First(&certdb)
-	}
-
-	Information.Entries = certdb
-
-	return Information, nil
-}
-
-func (c Cert) Paginated(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	var count int64
-	c.DB.Model(&Cert{}).Count(&count)
-	counter := int(count)
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		sql, err := vars.Sql(c)
-		if err != nil {
-			Information.Error = err.Error()
-			return Information, errors.New(dbError)
-		}
-		var certdb []Cert
-		c.DB.Select(sql.Select).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&certdb)
-		Information.Entries = certdb
-	}
-
-	return Information, nil
-}
-
-func (c Cert) Search(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	sql, err := vars.Sql(c)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-	var count int64
-	c.DB.Model(&Cert{}).Where(sql.Where.Query, sql.Where.Values...).Count(&count)
-	counter := int(count)
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		var certdb []Cert
-		c.DB.Select(sql.Select).Where(sql.Where.Query, sql.Where.Values...).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&certdb)
-		Information.Entries = certdb
-	}
-
-	return Information, nil
-}
-
-func (c Cert) Download(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	// Find the Cert
-	var cert Cert
-
-	if profile, ok := params["profile"]; ok {
-		if val, ok := params["cn"]; ok {
-			if CertDB := c.DB.Preload("Ca").Where("Cn = ? AND profile_id = ?", val, profile).Find(&cert); CertDB.Error != nil {
-				Information.Error = CertDB.Error.Error()
-				return Information, errors.New(dbError)
-			}
-		}
-		if val, ok := params["id"]; ok {
-			if CertDB := c.DB.Preload("Ca").Where("Id = ? AND profile_id = ?", val, profile).Find(&cert); CertDB.Error != nil {
-				Information.Error = CertDB.Error.Error()
-				return Information, errors.New(dbError)
-			}
-		}
-	} else {
-		if val, ok := params["id"]; ok {
-			if CertDB := c.DB.Preload("Ca").First(&cert, val); CertDB.Error != nil {
-				Information.Error = CertDB.Error.Error()
-				return Information, errors.New(dbError)
-			}
-		}
-	}
-
-	// Load the certificates from the database
-	certtls, err := tls.X509KeyPair([]byte(cert.Cert), []byte(cert.Key))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	// Find the profile
-	var prof Profile
-	if profDB := c.DB.Where("Name = ?", cert.ProfileName).Find(&prof); profDB.Error != nil {
-		Information.Error = profDB.Error.Error()
-		return Information, errors.New(dbError)
-	}
-
-	certificate, err := x509.ParseCertificate(certtls.Certificate[0])
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	// Load the certificates from the database
-	catls, err := tls.X509KeyPair([]byte(cert.Ca.Cert), []byte(cert.Ca.Key))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	var CaCert []*x509.Certificate
-
-	CaCert = append(CaCert, cacert)
-
-	var password string
-	if val, ok := params["password"]; ok {
-		password = val
-	} else {
-		password = certutils.GeneratePassword()
-	}
-	Information.Password = password
-
-	pkcs12, err := pkcs12.Encode(certutils.PRNG, certtls.PrivateKey, certificate, CaCert, password)
-
-	if _, ok := params["password"]; ok {
-		Information.Raw = pkcs12
-		Information.ContentType = "application/x-pkcs12"
-	} else {
-		Information, err = emailcert(c.Ctx, cert, prof, pkcs12, password)
-	}
-
-	return Information, err
-}
-
-func (c Cert) Revoke(params map[string]string) (types.Info, error) {
-
-	Information := types.Info{}
-	// Find the Cert
-	var cert Cert
-
-	reason := params["reason"]
-
-	if profile, ok := params["profile"]; ok {
-		if id, ok := params["id"]; ok {
-			if CertDB := c.DB.Preload("Ca").Where("id = ? AND profile_id = ?", id, profile).Find(&cert); CertDB.Error != nil {
-				Information.Error = CertDB.Error.Error()
-				return Information, CertDB.Error
-			}
-		}
-		if cn, ok := params["cn"]; ok {
-			if CertDB := c.DB.Preload("Ca").Where("cn = ? AND profile_id = ?", cn, profile).Find(&cert); CertDB.Error != nil {
-				Information.Error = CertDB.Error.Error()
-				return Information, CertDB.Error
-			}
-		}
-	} else {
-		if id, ok := params["id"]; ok {
-			if CertDB := c.DB.Preload("Ca").Where("id = ?", id).Find(&cert); CertDB.Error != nil {
-				Information.Error = CertDB.Error.Error()
-				return Information, CertDB.Error
-			}
-		}
-	}
-
-	// Find the Profile
-	var profile Profile
-
-	error := c.DB.Model(&profile).Where(&cert)
-	if error.Error != nil {
-		Information.Error = error.Error.Error()
-		return Information, error.Error
-	}
-
-	intreason, err := strconv.Atoi(reason)
-	if err != nil {
-		Information.Error = "Reason unsupported"
-		return Information, errors.New("Reason unsupported")
-	}
-	RevokeDate := time.Now().AddDate(0, 0, profile.RevokedValidUntil)
-	if err := c.DB.Create(&RevokedCert{Cn: cert.Cn, Mail: cert.Mail, Ca: cert.Ca, CaID: cert.CaID, CaName: cert.CaName, StreetAddress: cert.StreetAddress, Organisation: cert.Organisation, OrganisationalUnit: cert.OrganisationalUnit, Country: cert.Country, State: cert.State, Locality: cert.Locality, PostalCode: cert.Locality, Key: cert.Key, Cert: cert.Cert, Profile: profile, ProfileID: cert.ProfileID, ProfileName: cert.ProfileName, ValidUntil: cert.ValidUntil, NotBefore: cert.NotBefore, Date: cert.Date, Revoked: RevokeDate, CRLReason: intreason, SerialNumber: cert.SerialNumber, DNSNames: cert.DNSNames, IPAddresses: cert.IPAddresses, Subject: cert.Subject}).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	if err := c.DB.Unscoped().Delete(&cert).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	return Information, nil
-}
-
-func (c Cert) CheckRenewal(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var certdb []Cert
-
-	if CertDB := c.DB.Where("alert <> ?", 1).Find(&certdb); CertDB.Error != nil {
-		Information.Error = CertDB.Error.Error()
-		return Information, CertDB.Error
-	}
-
-	for _, v := range certdb {
-		// Find the profile
-		var prof Profile
-		if profDB := c.DB.First(&prof, v.ProfileID); profDB.Error != nil {
-			Information.Error = profDB.Error.Error()
-			return Information, errors.New(dbError)
-		}
-		// Revoke due certificate
-		if time.Now().Unix() > v.ValidUntil.Unix() {
-			params := make(map[string]string)
-
-			params["id"] = strconv.Itoa(int(v.ID))
-			params["reason"] = strconv.Itoa(ocsp.Superseded)
-			c.Revoke(params)
-		}
-		if prof.RenewalMail == 1 {
-			if *v.Scep == false {
-				if v.ValidUntil.Unix()-int64((time.Duration(prof.DaysBeforeRenewalMail)*24*time.Hour).Seconds()) < time.Now().Unix() {
-					emailRenewal(c.Ctx, v, prof)
-					notfalse := true
-					v.Alert = &notfalse
-					c.DB.Save(&v)
-				}
-			}
-		}
-	}
-
-	return Information, nil
-}
-
-func (c Cert) Resign(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var certdb []Cert
-	var err error
-	//Search the existing cert in the db
-	if val, ok := params["id"]; ok {
-		if err = c.DB.Preload("Ca").Preload("Profile").First(&certdb, val).Error; err != nil {
-			Information.Error = err.Error()
-			return Information, err
-		}
-	}
-
-	catls, err := tls.X509KeyPair([]byte(certdb[0].Ca.Cert), []byte(certdb[0].Ca.Key))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	Information.Entries = certdb
-
-	// Decode the private key
-	block, _ := pem.Decode([]byte(certdb[0].Key))
-	if block == nil {
-		log.LoggerWContext(c.Ctx).Error("failed to decode PEM block containing public key")
-	}
-
-	var skid []byte
-	var keyOut *bytes.Buffer
-	keyOut = new(bytes.Buffer)
-	var pub crypto.PublicKey
-
-	keyOut, skid, pub, _, Information, err = certutils.ExtractPrivateKey(certdb[0].Profile.KeyType, block, &Information)
-	if err != nil {
-		return Information, err
-	}
-
-	// keyOut contain the private key
-	var newcertdb []Cert
-
-	// Get serial number atomically using transaction with row-level locking
-	SerialNumber, err := getNextSerialNumber(c.DB, certdb[0].Ca.ID)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	Subject := certdb[0].MakeSubject()
-
-	cert := &x509.Certificate{
-		SerialNumber:       SerialNumber,
-		Subject:            Subject,
-		NotBefore:          time.Now(),
-		NotAfter:           time.Now().AddDate(0, 0, certdb[0].Profile.Validity),
-		SignatureAlgorithm: certutils.CompatibleSigAlgo(*certdb[0].Ca.KeyType, certdb[0].Profile.Digest),
-		ExtKeyUsage:        certutils.Extkeyusage(strings.Split(*certdb[0].Profile.ExtendedKeyUsage, "|")),
-		KeyUsage:           x509.KeyUsage(certutils.Keyusage(strings.Split(*certdb[0].Profile.KeyUsage, "|"))),
-		SubjectKeyId:       skid,
-	}
-	//Overload certificate attributes
-	if len(c.Profile.OCSPUrl) > 0 {
-		cert.OCSPServer = []string{c.Profile.OCSPUrl}
-	}
-
-	Email := ""
-	if len(certdb[0].Profile.Mail) > 0 {
-		Email = certdb[0].Profile.Mail
-	}
-	if len(c.Mail) > 0 {
-		Email = c.Mail
-	}
-	if len(Email) > 0 {
-		for _, mail := range strings.Split(Email, ",") {
-			cert.EmailAddresses = append(cert.EmailAddresses, mail)
-		}
-	}
-
-	if len(c.DNSNames) > 0 {
-		for _, dns := range strings.Split(c.DNSNames, ",") {
-			cert.DNSNames = append(cert.DNSNames, dns)
-		}
-	}
-	var IPAddresses []string
-	if len(c.IPAddresses) > 0 {
-		for _, ip := range strings.Split(c.IPAddresses, ",") {
-			if net.ParseIP(ip) == nil {
-				fmt.Printf("IP Address: %s - Invalid\n", ip)
-			} else {
-				IPAddresses = append(IPAddresses, ip)
-				cert.IPAddresses = append(cert.IPAddresses, net.ParseIP(ip))
-			}
-		}
-	}
-
-	var certBytes []byte
-
-	switch *certdb[0].Profile.KeyType {
-	case certutils.KEY_RSA:
-		certBytes, err = x509.CreateCertificate(rand.Reader, cert, cacert, pub, catls.PrivateKey.(*rsa.PrivateKey))
-	case certutils.KEY_ECDSA:
-		certBytes, err = x509.CreateCertificate(rand.Reader, cert, cacert, pub, catls.PrivateKey.(*ecdsa.PrivateKey))
-	case certutils.KEY_DSA:
-		certBytes, err = x509.CreateCertificate(rand.Reader, cert, cacert, pub, catls.PrivateKey.(*dsa.PrivateKey))
-	case certutils.KEY_ED25519:
-		certBytes, err = x509.CreateCertificate(rand.Reader, cert, cacert, pub, catls.PrivateKey.(ed25519.PrivateKey))
-	}
-	if err != nil {
-		return Information, err
-	}
-
-	certBuff := new(bytes.Buffer)
-	// Public key
-	pem.Encode(certBuff, &pem.Block{Type: "CERTIFICATE", Bytes: certBytes})
-
-	h := sha1.New()
-
-	h.Write(cacert.RawIssuer)
-	if err := c.DB.Model(&Cert{}).Where("cn = ?", c.Cn).Updates(map[string]interface{}{"Cn": c.Cn, "Ca": certdb[0].Ca, "CaName": certdb[0].Ca.Cn, "ProfileName": certdb[0].Profile.Name, "SerialNumber": SerialNumber.String(), "DNSNames": cert.DNSNames, "IPAddresses": strings.Join(IPAddresses, ","), "Mail": Email, "StreetAddress": cert.Subject.StreetAddress, "Organisation": cert.Subject.Organization, "OrganisationalUnit": cert.Subject.OrganizationalUnit, "Country": cert.Subject.Country, "State": cert.Subject.Province, "Locality": cert.Subject.Locality, "PostalCode": cert.Subject.PostalCode, "Profile": certdb[0].Profile, "Key": keyOut.String(), "Cert": certBuff.String(), "ValidUntil": cert.NotAfter, "NotBefore": cert.NotBefore, "Subject": cert.Subject.String()}).Error; err != nil {
-		Information.Error = err.Error()
-		Information.Status = http.StatusConflict
-		return Information, errors.New(dbError)
-	}
-
-	c.DB.Where("cn = ? AND profile_name = ?", c.Cn, certdb[0].ProfileName).First(&newcertdb)
-	Information.Entries = newcertdb
-	Information.Serial = SerialNumber.String()
-
-	return Information, nil
-}
-
-func (c Cert) MakeSubject() pkix.Name {
-	var Subject pkix.Name
-	Subject.CommonName = c.Cn
-
-	//Overload certificate attributes if exist
-	Organization := ""
-	if len(c.Profile.Organisation) > 0 {
-		Organization = c.Profile.Organisation
-	}
-	if len(c.Organisation) > 0 {
-		Organization = c.Organisation
-	}
-	if len(Organization) > 0 {
-		Subject.Organization = []string{Organization}
-	}
-
-	Country := ""
-	if len(c.Profile.Country) > 0 {
-		Country = c.Profile.Country
-	}
-	if len(c.Country) > 0 {
-		Country = c.Country
-	}
-	if len(Country) > 0 {
-		Subject.Country = []string{Country}
-	}
-
-	Province := ""
-	if len(c.Profile.State) > 0 {
-		Province = c.Profile.State
-	}
-	if len(c.State) > 0 {
-		Province = c.State
-	}
-	if len(Province) > 0 {
-		Subject.Province = []string{Province}
-	}
-
-	Locality := ""
-	if len(c.Profile.Locality) > 0 {
-		Locality = c.Profile.Locality
-	}
-	if len(c.Locality) > 0 {
-		Locality = c.Locality
-	}
-	if len(Locality) > 0 {
-		Subject.Locality = []string{Locality}
-	}
-
-	StreetAddress := ""
-	if len(c.Profile.StreetAddress) > 0 {
-		StreetAddress = c.Profile.StreetAddress
-	}
-	if len(c.StreetAddress) > 0 {
-		StreetAddress = c.StreetAddress
-	}
-	if len(StreetAddress) > 0 {
-		Subject.StreetAddress = []string{StreetAddress}
-	}
-
-	PostalCode := ""
-	if len(c.Profile.PostalCode) > 0 {
-		PostalCode = c.Profile.PostalCode
-	}
-	if len(c.PostalCode) > 0 {
-		PostalCode = c.PostalCode
-	}
-	if len(PostalCode) > 0 {
-		Subject.PostalCode = []string{PostalCode}
-	}
-	return Subject
-}
-
-func NewRevokedCertModel(pfpki *types.Handler) *RevokedCert {
-	RevokedCert := &RevokedCert{}
-
-	RevokedCert.DB = pfpki.DB
-	RevokedCert.Ctx = *pfpki.Ctx
-
-	return RevokedCert
-}
-
-func (c RevokedCert) GetByID(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var revokedcertdb []RevokedCert
-	if val, ok := params["id"]; ok {
-		allFields := strings.Join(sql.SqlFields(c)[:], ",")
-		c.DB.Select(allFields).Where("`id` = ?", val).First(&revokedcertdb)
-	}
-	Information.Entries = revokedcertdb
-
-	return Information, nil
-}
-
-func (c RevokedCert) Paginated(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	var count int64
-	c.DB.Model(&RevokedCert{}).Count(&count)
-	counter := int(count)
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		sql, err := vars.Sql(c)
-		if err != nil {
-			Information.Error = err.Error()
-			return Information, err
-		}
-		var revokedcertdb []RevokedCert
-		result := c.DB.Select(sql.Select).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&revokedcertdb)
-		if result.Error != nil {
-			Information.Error = result.Error.Error()
-			return Information, err
-		}
-
-		Information.Entries = revokedcertdb
-	}
-
-	return Information, nil
-}
-
-func (c RevokedCert) Search(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	sql, err := vars.Sql(c)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	var count int64
-	c.DB.Model(&Cert{}).Where(sql.Where.Query, sql.Where.Values...).Count(&count)
-	counter := int(count)
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		var revokedcertdb []RevokedCert
-		c.DB.Select(sql.Select).Where(sql.Where.Query, sql.Where.Values...).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&revokedcertdb)
-		Information.Entries = revokedcertdb
-	}
-
-	return Information, nil
-}
-
-func NewCsrModel(pfpki *types.Handler) *CSR {
-	Csr := &CSR{}
-
-	Csr.DB = pfpki.DB
-	Csr.Ctx = *pfpki.Ctx
-
-	return Csr
-}
-
-func (csr CSR) New(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	Information.Status = http.StatusUnprocessableEntity
-	// Find the profile
-	var prof Profile
-	if val, ok := params["id"]; ok {
-
-		if profDB := csr.DB.First(&prof, val); profDB.Error != nil {
-			Information.Error = profDB.Error.Error()
-			return Information, errors.New(dbError)
-		}
-	} else {
-		return Information, errors.New("Missing the profile id in the url")
-
-	}
-	attributes := ProfileAttributes(prof)
-	// Find the CA
-	var ca CA
-	if err := csr.DB.First(&ca, prof.CaID).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-
-	// Decode the CSR. Accept PEM (with BEGIN/END markers) or raw DER.
-	var (
-		err error
-		der []byte
-	)
-	if block, _ := pem.Decode([]byte(csr.Csr)); block != nil {
-		der = block.Bytes
-	} else {
-		der = []byte(csr.Csr)
-	}
-
-	certRequest, err := x509.ParseCertificateRequest(der)
-	if err != nil {
-		Information.Error = err.Error()
-		Information.Status = http.StatusUnprocessableEntity
-		return Information, err
-	}
-
-	id, err := cryptoutil.GenerateSubjectKeyID(certRequest.PublicKey)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	ca.DB = csr.DB
-
-	serial, err := ca.FindSerial(prof)
-
-	if err != nil {
-		return Information, err
-	}
-
-	Subject := certutils.MakeSubject(certRequest.Subject, attributes)
-	Subject.CommonName = certRequest.Subject.CommonName
-
-	ExtKeyUsage := certutils.Extkeyusage(strings.Split(attributes["ExtendedKeyUsage"], "|"))
-	KeyUsage := x509.KeyUsage(certutils.Keyusage(strings.Split(attributes["KeyUsage"], "|")))
-
-	// create cert template
-	v, _ := strconv.Atoi(attributes["Digest"])
-	SignatureAlgorithm := certutils.CompatibleSigAlgo(*ca.KeyType, x509.SignatureAlgorithm(v))
-
-	var ExtraExtensions []pkix.Extension
-
-	for _, v := range certRequest.Extensions {
-		if v.Id.String() != "2.5.29.37" {
-			if v.Id.String() == "2.5.29.17" {
-				ext, err := certutils.ForEachSAN(v.Value, attributes)
-				if err == nil {
-					ExtraExtensions = append(ExtraExtensions, ext)
-				}
-			} else {
-				ExtraExtensions = append(ExtraExtensions, v)
-			}
-		}
-
-	}
-
-	tmpl := &x509.Certificate{
-		SerialNumber:       serial,
-		Subject:            Subject,
-		NotBefore:          time.Now().Add(-600).UTC(),
-		NotAfter:           time.Now().AddDate(0, 0, prof.Validity).UTC(),
-		SubjectKeyId:       id,
-		KeyUsage:           KeyUsage,
-		ExtKeyUsage:        ExtKeyUsage,
-		SignatureAlgorithm: SignatureAlgorithm,
-		DNSNames:           certRequest.DNSNames,
-		EmailAddresses:     certRequest.EmailAddresses,
-		IPAddresses:        certRequest.IPAddresses,
-		URIs:               certRequest.URIs,
-		ExtraExtensions:    ExtraExtensions,
-	}
-
-	if len(attributes["OCSPUrl"]) > 0 {
-		tmpl.OCSPServer = []string{attributes["OCSPUrl"]}
-	}
-
-	if len(attributes["Mail"]) > 0 {
-		tmpl.EmailAddresses = []string{attributes["Mail"]}
-	}
-
-	// Load the certificates from the database
-	catls, err := tls.X509KeyPair([]byte(ca.Cert), []byte(ca.Key))
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	cacert, err := x509.ParseCertificate(catls.Certificate[0])
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	// Sign the certificate
-	certByte, err := x509.CreateCertificate(rand.Reader, tmpl, cacert, certRequest.PublicKey, catls.PrivateKey)
-
-	certBuff := new(bytes.Buffer)
-
-	// Public key
-	pem.Encode(certBuff, &pem.Block{Type: "CERTIFICATE", Bytes: certByte})
-
-	c := &Cert{}
-
-	c.DB = csr.DB
-	c.Ctx = csr.Ctx
-
-	var IPAddresses []string
-	for _, IP := range certRequest.IPAddresses {
-		IPAddresses = append(IPAddresses, IP.String())
-	}
-
-	attributeMap := certutils.GetDNFromCert(Subject)
-
-	certif, err := x509.ParseCertificate(certByte)
-	name := certutils.CertName(certif)
-	notfalse := true
-
-	newCert := Cert{Cn: name, Ca: ca, CaName: ca.Cn, ProfileName: prof.Name, SerialNumber: serial.String(), DNSNames: c.DNSNames, IPAddresses: strings.Join(IPAddresses, ","), Mail: strings.Join(certRequest.EmailAddresses, ","), StreetAddress: attributeMap["streetAddress"], Organisation: attributeMap["O"], OrganisationalUnit: attributeMap["OU"], Country: attributeMap["C"], State: attributeMap["ST"], Locality: attributeMap["L"], PostalCode: attributeMap["postalCode"], Profile: prof, Cert: certBuff.String(), ValidUntil: tmpl.NotAfter, Subject: Subject.String(), Csr: &notfalse}
-	if err := c.DB.Create(&newCert).Error; err != nil {
-		Information.Error = err.Error()
-		Information.Status = http.StatusConflict
-		return Information, errors.New(dbError)
-	}
-	Information.Entries = []Cert{newCert}
-	Information.Serial = serial.String()
-
-	return Information, nil
-
-}
-
-// EmailType strucure
-type EmailType struct {
-	Header   string
-	Footer   string
-	Password string
-	To       string
-	From     string
-	Subject  string
-	FileName string
-	File     []byte
-	Template string
-}
-
-func emailcert(ctx context.Context, cert Cert, profile Profile, file []byte, password string) (types.Info, error) {
-	alerting := pfconfigdriver.GetType[pfconfigdriver.PfConfAlerting](ctx)
-
-	mail := EmailType{Header: profile.P12MailHeader, Footer: profile.P12MailFooter}
-	if len(profile.P12MailFrom) > 0 {
-		mail.From = profile.P12MailFrom
-	} else if len(alerting.FromAddr) > 0 {
-		mail.From = alerting.FromAddr
-	} else {
-		name, _ := os.Hostname()
-		mail.From = "root@" + name
-	}
-	mail.To = cert.Mail
-	mail.Subject = profile.P12MailSubject
-	mail.FileName = cert.Cn
-	mail.Template = "emails-pki_certificate.html"
-	mail.File = file
-	mail.Password = password
-	return email(ctx, mail)
-}
-
-func emailRenewal(ctx context.Context, cert Cert, profile Profile) (types.Info, error) {
-	alerting := pfconfigdriver.GetType[pfconfigdriver.PfConfAlerting](ctx)
-
-	mail := EmailType{}
-
-	if len(profile.RenewalMailFrom) > 0 {
-		mail.From = profile.RenewalMailFrom
-	} else {
-		mail.From = alerting.FromAddr
-	}
-
-	if len(cert.Mail) > 0 {
-		mail.To = cert.Mail
-	} else if len(profile.Mail) > 0 {
-		mail.To = profile.Mail
-	} else if len(alerting.EmailAddr) > 0 {
-		mail.To = alerting.EmailAddr
-	} else {
-		name, _ := os.Hostname()
-		mail.From = "root@" + name
-	}
-	mail.Subject = profile.RenewalMailSubject
-	mail.FileName = "Profile Name: " + profile.Name + " Certificate CN: " + cert.Cn
-	mail.Template = "emails-renewal_certificate.html"
-	mail.Header = profile.RenewalMailHeader
-	mail.Footer = profile.RenewalMailFooter
-
-	return email(ctx, mail)
-}
-
-func email(ctx context.Context, email EmailType) (types.Info, error) {
-	alerting := pfconfigdriver.GetType[pfconfigdriver.PfConfAlerting](ctx)
-	advanced := pfconfigdriver.GetType[pfconfigdriver.PfConfAdvanced](ctx)
-
-	Information := types.Info{}
-
-	dict, err := ParseYAMLDict()
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	cat, err := catalog.NewFromMap(dict)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-	message.DefaultCatalog = cat
-
-	m := gomail.NewMessage()
-
-	m.SetHeader("From", email.From)
-	m.SetHeader("To", email.To)
-	m.SetHeader("Subject", email.Subject)
-
-	lang := language.MustParse(advanced.Language)
-
-	emailContent, err := parseTemplate(email.Template, lang, email)
-
-	m.SetBody("text/html", emailContent)
-	if len(email.File) > 0 {
-		m.Attach(email.FileName+".p12", gomail.SetCopyFunc(func(w io.Writer) error {
-			_, err := w.Write(email.File)
-			return err
-		}))
-	}
-	d := gomail.NewDialer(alerting.SMTPServer, alerting.SMTPPort, alerting.SMTPUsername, alerting.SMTPPassword.String())
-
-	if alerting.SMTPVerifySSL == "disabled" || alerting.SMTPEncryption == "none" {
-		d.TLSConfig = &tls.Config{InsecureSkipVerify: true}
-	}
-
-	if err := d.DialAndSend(m); err != nil {
-		Information.Error = err.Error()
-		return Information, err
-	}
-
-	return Information, nil
-}
-
-func parseTemplate(tplName string, lang language.Tag, data interface{}) (string, error) {
-	p := message.NewPrinter(lang)
-	fmap := template.FuncMap{
-		"translate": p.Sprintf,
-	}
-
-	t, err := template.New(tplName).Funcs(fmap).ParseFiles("/usr/local/pf/html/captive-portal/templates/emails/" + tplName)
-	if err != nil {
-		return "", err
-	}
-
-	buf := bytes.NewBuffer([]byte{})
-	if err := t.Execute(buf, data); err != nil {
-		return "", err
-	}
-
-	return buf.String(), nil
-}
-
-func ParseYAMLDict() (map[string]catalog.Dictionary, error) {
-	dir := "/usr/local/pf/conf/caddy-services/locales"
-	files, err := os.ReadDir(dir)
-	if err != nil {
-		fmt.Println(err)
-		return nil, err
-	}
-
-	translations := map[string]catalog.Dictionary{}
-
-	for _, f := range files {
-		yamlFile, err := os.ReadFile(dir + "/" + f.Name())
-		if err != nil {
-			return nil, err
-		}
-		data := map[string]string{}
-		err = yaml.Unmarshal(yamlFile, &data)
-		if err != nil {
-			return nil, err
-		}
-
-		lang := strings.Split(f.Name(), ".")[0]
-
-		translations[lang] = &dictionary{Data: data}
-	}
-
-	return translations, nil
-}
-
-type dictionary struct {
-	Data map[string]string
-}
-
-func (d *dictionary) Lookup(key string) (data string, ok bool) {
-	if _, ok := d.Data[key]; !ok {
-		return "", false
-	}
-
-	return "\x02" + d.Data[key], true
-}
 
 func ProfileAttributes(prof Profile) map[string]string {
 	var attributes map[string]string
@@ -2241,109 +322,3 @@ func ProfileAttributes(prof Profile) map[string]string {
 	return attributes
 }
 
-func NewSCEPServerModel(pfpki *types.Handler) *SCEPServer {
-	SCEPServer := &SCEPServer{}
-
-	SCEPServer.DB = pfpki.DB
-	SCEPServer.Ctx = *pfpki.Ctx
-
-	return SCEPServer
-}
-
-func (s SCEPServer) New() (types.Info, error) {
-	Information := types.Info{}
-
-	newSCEP := SCEPServer{Name: s.Name, URL: s.URL, SharedSecret: s.SharedSecret}
-	if err := s.DB.Create(&newSCEP).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-	Information.Entries = []SCEPServer{newSCEP}
-
-	return Information, nil
-}
-
-// GetByID retreive the SCEPServer by id
-func (s SCEPServer) GetByID(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var scepserverdb []SCEPServer
-	if val, ok := params["id"]; ok {
-		allFields := strings.Join(sql.SqlFields(s)[:], ",")
-		s.DB.Select(allFields).Where("`id` = ?", val).First(&scepserverdb)
-	}
-	Information.Entries = scepserverdb
-
-	return Information, nil
-}
-
-// GetByID retreive the SCEPServer by id
-func (s SCEPServer) DelByID(params map[string]string) (types.Info, error) {
-	Information := types.Info{}
-	var scepserverdb []SCEPServer
-	if val, ok := params["id"]; ok {
-		s.DB.Delete(&SCEPServer{}, val)
-	}
-	Information.Entries = scepserverdb
-	return Information, nil
-}
-
-// Search for the SCEPServer
-func (s SCEPServer) Search(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	sql, err := vars.Sql(s)
-	if err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-	var count int64
-	s.DB.Model(&SCEPServer{}).Where(sql.Where.Query, sql.Where.Values...).Count(&count)
-	counter := int(count)
-
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		var scepserverdb []SCEPServer
-		s.DB.Select(sql.Select).Where(sql.Where.Query, sql.Where.Values...).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&scepserverdb)
-		Information.Entries = scepserverdb
-	}
-
-	return Information, nil
-}
-
-func (s SCEPServer) Update() (types.Info, error) {
-	Information := types.Info{}
-	if err := s.DB.Model(&SCEPServer{}).Where("name = ?", s.Name).Updates(map[string]interface{}{"url": s.URL, "shared_secret": s.SharedSecret}).Error; err != nil {
-		Information.Error = err.Error()
-		return Information, errors.New(dbError)
-	}
-	var scepserverdb []SCEPServer
-	s.DB.Where("name = ?", s.Name).First(&scepserverdb)
-	Information.Entries = scepserverdb
-
-	return Information, nil
-}
-
-// Paginated return the SCEPServer list paginated
-func (s SCEPServer) Paginated(vars sql.Vars) (types.Info, error) {
-	Information := types.Info{}
-	var count int64
-	s.DB.Model(&CA{}).Count(&count)
-	counter := int(count)
-
-	Information.TotalCount = counter
-	Information.PrevCursor = vars.Cursor
-	Information.NextCursor = vars.Cursor + vars.Limit
-	if vars.Cursor < counter {
-		sql, err := vars.Sql(s)
-		if err != nil {
-			Information.Error = err.Error()
-			return Information, errors.New(dbError)
-		}
-		var scepserverdb []SCEPServer
-		s.DB.Select(sql.Select).Order(sql.Order).Offset(sql.Offset).Limit(sql.Limit).Find(&scepserverdb)
-		Information.Entries = scepserverdb
-	}
-
-	return Information, nil
-}
