+++ go/plugin/caddy2/pfpki/pfpki.go
@@ -104,6 +104,7 @@ func (h *Handler) buildPfpkiHandler(ctx context.Context) error {
 	r.Use(middleware.RealIP)
 	r.Use(middleware.Logger)
 	r.Use(middleware.Recoverer)
+	r.Use(handlers.LimitRequestBody)
 
 	PFPki := &pfpki
